1st Attempt at User Input
Added Attacks collapsable table and tried to play around with user input, not entirely successful. I also tried adding an ATTACK button for the fun of it since I know it'll be required later, sadly that also isn't working at the moment. 
With the user input, when I tried specifying where to print the text (ie, under the Name column in the Monsters table), I got errors. I've decided to leave it this way for now and hopefully I can fix things when I get back or someone will gleam some insight in the meantime.